build: ensure image is built by default
We found after the repository went public that docker images are not
able to hosted publicly within the organization. This caused build
breaks since outside users were not able to pull the image and also
prevents CI from running in forks.

To restore functionality:

- add logic to ensure the image is built if it cannot be found locally
- in CI, add logic to see if the image cannot be pulled, indicating a
  fork, and if so, build the image accordingly

Additionally, the image repository name is duplicated in many places. To
reduce this duplication, have one source of truth for this value within
`docker/setupDockerEnv` to allow for sourcing by the various Docker
processes.

Refs: BARTON-250